add keys by name
this enables -s/-l to use key name, and prints key names in event
messages

This grows the binary a bit:
32K	buttond-before
56K	buttond-after

   text	   data	    bss	    dec	    hex	filename
  18824	   1160	     64	  20048	   4e50	buttond-before
  36504	   7312	     64	  43880	   ab68	buttond-after

which is not surprising given how many strings we include in there.
If size turns out to be a problem, we can trim it down a bit to
only include useful names? or store names in a trie that could be more
efficient