time_diff*: round up to next ms
For short timeouts we would sometimes be woken up before target date,
leading to 'woke up but no action found' error.

When we should wake up after 9.5ms make sure we sleep 10ms and not 9ms
to fix that.

Note that it's also possible for poll to just wake up early for no
reason (e.g. EINTR), in that case we'll just setup a new poll with new
time diff, so this fix should be complete with just that.

Change-Id: I3c2237889e23de5c241df5e85a5e3c0b8ef8071b